+++ edgelet/aziot-edged/src/main.rs
@@ -13,8 +13,8 @@ mod workload_manager;
 use std::sync::atomic;
 
 use edgelet_core::{module::ModuleAction, MakeModuleRuntime, ModuleRuntime};
+use edgelet_docker::MIGCPersistence;
 use edgelet_settings::RuntimeSettings;
-use edgelet_utils::migc_persistence::MIGCPersistence;
 
 use crate::{error::Error as EdgedError, workload_manager::WorkloadManager};
