+++ edgelet/aziot-edged/src/management.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 
+use edgelet_docker::MIGCPersistence;
 use edgelet_settings::uri::Listen;
-use edgelet_utils::migc_persistence::MIGCPersistence;
 
 use crate::error::Error as EdgedError;
