+++ edgelet/edgelet-docker/src/lib.rs
@@ -8,14 +8,18 @@
     clippy::module_name_repetitions,
     clippy::must_use_candidate,
     clippy::too_many_lines,
-    clippy::use_self
+    clippy::use_self,
+    // TODO: below this revert
+    dead_code,
 )]
 
 // mod client;
 mod error;
+mod migc_persistence;
 mod module;
 mod runtime;
 
 pub use error::Error;
+pub use migc_persistence::MIGCPersistence;
 pub use module::{DockerModule, MODULE_TYPE};
 pub use runtime::{init_client, DockerModuleRuntime};
