+++ edgelet/edgelet-http-mgmt/Cargo.toml
@@ -20,6 +20,7 @@ tokio = { version = "1", features = ["macros", "sync"] }
 url = "2"
 
 edgelet-core = { path = "../edgelet-core" }
+edgelet-docker = { path = "../edgelet-docker" }
 edgelet-http = { path = "../edgelet-http" }
 edgelet-settings = { path = "../edgelet-settings" }
 edgelet-utils = { path = "../edgelet-utils" }
