--- edgelet/edgelet-utils/src/migc_persistence.rs
@@ -1,12 +1,0 @@
-#[derive(Debug, Clone)]
-pub struct MIGCPersistence {
-    _data: String,
-}
-
-impl MIGCPersistence {
-    pub fn new() -> Self {}
-
-    pub fn write_image_to_file() {}
-
-    pub fn prune_images_from_file();
-}
